cloud-requirements.txt
Django==1.11.17
django-filter==1.1.0
djangorestframework==3.8.2
djangorestframework-filters==0.10.2
django-haystack
mysqlclient==1.3.13
pytz==2018.5
pysolr
certifi==2018.8.24
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
idna==2.7
itypes==1.1.0
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
Pygments==2.2.0
requests==2.19.1
django-cors-headers==2.4.0
